requirements.txt
alembic==1.6.5
certifi==2021.5.30
chardet==4.0.0
Click==7.0
Flask==1.1.1
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.4
MarkupSafe==1.1.1
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.11.0
python-editor==1.0.4
requests==2.25.1
six==1.16.0
SQLAlchemy==1.4.18
urllib3==1.26.5
Werkzeug==1.0.0
